autovid/requirements.txt
# ОСНОВНЫЕ
prefect>=2.14.0
celery>=5.3.0
redis>=4.6.0
sqlalchemy>=2.0.0
ffmpeg>=6.0.0

# AutoVid core
openai>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
mutagen>=1.45.1
python-multipart>=0.0.14

# Новые для ускоренной генерации
opencv-python-headless>=4.10.0     # или opencv-python
tqdm>=4.66.0

# Инфраструктура/утилиты
psutil>=5.9.0
watchdog>=3.0.0
aiohttp>=3.10.0
pydub>=0.25.1
requests-toolbelt>=0.10.1
google-api-python-client>=2.15.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.1.0
